requirements.txt
discord.py
discord.py[voice]
python-dotenv
ffmpeg
ffprobe
youtube-dl
wikipedia
googletrans==3.1.0a0
requests
bs4
regex